+++ test/test.sql
@@ -113,4 +113,6 @@ CREATE USER test_cartodb_user_1 WITH PASSWORD '';
 GRANT ALL ON TABLE untitle_table_4 TO test_cartodb_user_1;
 GRANT SELECT ON TABLE untitle_table_4 TO publicuser;
 GRANT ALL ON TABLE private_table TO test_cartodb_user_1;
-GRANT ALL ON SEQUENCE test_table_cartodb_id_seq_p TO test_cartodb_user_1
+GRANT ALL ON SEQUENCE test_table_cartodb_id_seq_p TO test_cartodb_user_1;
+
+GRANT ALL ON TABLE spatial_ref_sys TO test_cartodb_user_1, publicuser;